#80 [REFACTOR] 외부 서비스 mapper를 infra 패키지로 이전
