+++ src/main/java/com/wakeUpTogetUp/togetUp/infra/azure/vision/AzureVision40Service.java
@@ -13,8 +13,8 @@ import com.azure.ai.vision.imageanalysis.ImageAnalysisResultDetails;
 import com.azure.ai.vision.imageanalysis.ImageAnalysisResultReason;
 import com.azure.ai.vision.imageanalysis.ImageAnalyzer;
 import com.wakeUpTogetUp.togetUp.api.mission.domain.ObjectDetectionResult;
-import com.wakeUpTogetUp.togetUp.api.mission.service.mapper.ObjectDetectedV40Mapper;
-import com.wakeUpTogetUp.togetUp.api.mission.service.mapper.TagDetectedV40Mapper;
+import com.wakeUpTogetUp.togetUp.infra.azure.vision.mapper.ObjectDetectedV40Mapper;
+import com.wakeUpTogetUp.togetUp.infra.azure.vision.mapper.TagDetectedV40Mapper;
 import com.wakeUpTogetUp.togetUp.common.Status;
 import com.wakeUpTogetUp.togetUp.exception.BaseException;
 import java.io.IOException;
